requirements.txt
requests
bs4
yfinance
googlesearch-python
selenium
pillow
python-dotenv
cloudflare-functions
beautifulsoup4
langchain
